pyproject.toml
[project]
name = "smartbot"
version = "0.1.0"
description = "Modular and extensible chatbot core with pluggable providers and memory backends"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "pyyaml>=6.0.3",
    "pydantic>=2.12.5",
    "openai>=2.20.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
    "pyright>=1.1.408",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# =============================================================================
# RUFF — Linter y formateador
# =============================================================================

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "W", "F",
    "N", "I", "UP",
    "C90", "B", "SIM", "C4", "RET", "PIE",
    "T20", "ERA", "ARG",
    "PLR", "PLW",
    "PTH", "RUF",
]

ignore = [
    "PLR2004",
    "PLR0913",
    "ARG002",
    "RET504",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ARG",
    "PLR2004",
    "T20",
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.lint.pylint]
max-args = 5
max-returns = 4
max-branches = 8

[tool.ruff.lint.isort]
known-first-party = ["smartbot"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# =============================================================================
# PYRIGHT — Type checking
# =============================================================================

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
venvPath = "."
venv = ".venv"

include = ["src", "tests"]

reportMissingTypeStubs = false
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportMissingParameterType = "warning"
reportUnnecessaryIsInstance = "warning"
